data handle use stratgy pattern
