+++ tdenginewriter/src/main/java/com/alibaba/datax/plugin/writer/DataHandler.java
@@ -1,0 +1,10 @@
+package com.alibaba.datax.plugin.writer;
+
+import com.alibaba.datax.common.plugin.RecordReceiver;
+
+import java.util.Properties;
+
+public interface DataHandler {
+
+    long handle(RecordReceiver lineReceiver, Properties properties);
+}
