+++ tdenginewriter/src/main/java/com/alibaba/datax/plugin/writer/DataHandlerFactory.java
@@ -1,0 +1,10 @@
+package com.alibaba.datax.plugin.writer;
+
+public class DataHandlerFactory {
+
+    public static DataHandler build(String peerPluginName) {
+        if (peerPluginName.equals("opentsdbreader"))
+            return new OpentsdbDataHandler();
+        return new DefaultDataHandler();
+    }
+}
