+++ tdenginewriter/src/main/java/com/alibaba/datax/plugin/writer/TDengineWriter.java
@@ -1,21 +1,20 @@
 package com.alibaba.datax.plugin.writer;
 
 
-import com.alibaba.datax.common.element.Column;
-import com.alibaba.datax.common.element.Record;
-import com.alibaba.datax.common.exception.DataXException;
 import com.alibaba.datax.common.plugin.RecordReceiver;
 import com.alibaba.datax.common.spi.Writer;
 import com.alibaba.datax.common.util.Configuration;
 import org.slf4j.Logger;
 import org.slf4j.LoggerFactory;
 
-import java.util.*;
+import java.util.ArrayList;
+import java.util.List;
+import java.util.Properties;
+import java.util.Set;
 
 public class TDengineWriter extends Writer {
 
     private static final String PEER_PLUGIN_NAME = "peerPluginName";
-    private static final String DEFAULT_BATCH_SIZE = "1";
 
     public static class Job extends Writer.Job {
 
@@ -67,85 +66,11 @@ public class TDengineWriter extends Writer {
             }
 
             String peerPluginName = this.writerSliceConfig.getString(PEER_PLUGIN_NAME);
-            if (peerPluginName.equals("opentsdbreader")) {
-                // opentsdb json protocol use JNI and schemaless API to write
-                String host = properties.getProperty(Key.HOST);
-                int port = Integer.parseInt(properties.getProperty(Key.PORT));
-                String dbname = properties.getProperty(Key.DBNAME);
-                String user = properties.getProperty(Key.USER);
-                String password = properties.getProperty(Key.PASSWORD);
-
-                JniConnection conn = null;
-                try {
-                    conn = new JniConnection(properties);
-                    conn.open(host, port, dbname, user, password);
-                    LOG.info("TDengine connection established, host: " + host + ", port: " + port + ", dbname: " + dbname + ", user: " + user);
-                    int batchSize = Integer.parseInt(properties.getProperty(Key.BATCH_SIZE, DEFAULT_BATCH_SIZE));
-                    writeOpentsdb(lineReceiver, conn, batchSize);
-                } catch (Exception e) {
-                    LOG.error(e.getMessage());
-                    e.printStackTrace();
-                } finally {
-                    try {
-                        if (conn != null)
-                            conn.close();
-                    } catch (Exception e) {
-                        e.printStackTrace();
-                    }
-                    LOG.info("TDengine connection closed");
-                }
-            } else {
-                // other
-            }
+            LOG.debug("start to handle record from: " + peerPluginName);
+            DataHandler handler = DataHandlerFactory.build(peerPluginName);
+            long records = handler.handle(lineReceiver, properties);
+            LOG.debug("handle data finished, records: " + records);
         }
 
-        private void writeOpentsdb(RecordReceiver lineReceiver, JniConnection conn, int batchSize) {
-            try {
-                Record record;
-                StringBuilder sb = new StringBuilder();
-                long recordIndex = 1;
-                while ((record = lineReceiver.getFromReader()) != null) {
-                    if (batchSize == 1) {
-                        String jsonData = recordToString(record);
-                        LOG.debug(">>> " + jsonData);
-                        conn.insertOpentsdbJson(jsonData);
-                    } else if (recordIndex % batchSize == 1) {
-                        sb.append("[").append(recordToString(record)).append(",");
-                    } else if (recordIndex % batchSize == 0) {
-                        sb.append(recordToString(record)).append("]");
-                        String jsonData = sb.toString();
-                        LOG.debug(">>> " + jsonData);
-                        conn.insertOpentsdbJson(jsonData);
-                        sb.delete(0, sb.length());
-                    } else {
-                        sb.append(recordToString(record)).append(",");
-                    }
-                    recordIndex++;
-                }
-                if (sb.length() != 0 && sb.charAt(0) == '[') {
-                    String jsonData = sb.deleteCharAt(sb.length() - 1).append("]").toString();
-                    LOG.debug(">>> " + jsonData);
-                    conn.insertOpentsdbJson(jsonData);
-                }
-            } catch (Exception e) {
-                LOG.error("TDengineWriter ERROR: " + e.getMessage());
-                throw DataXException.asDataXException(TDengineWriterErrorCode.RUNTIME_EXCEPTION, e);
-            }
-        }
-
-        private String recordToString(Record record) {
-            int recordLength = record.getColumnNumber();
-            if (0 == recordLength) {
-                return "";
-            }
-            Column column;
-            StringBuilder sb = new StringBuilder();
-            for (int i = 0; i < recordLength; i++) {
-                column = record.getColumn(i);
-                sb.append(column.asString()).append("\t");
-            }
-            sb.setLength(sb.length() - 1);
-            return sb.toString();
-        }
     }
 }
